Refactor monumentsControler, abstract handler at a class
