+++ src/app/app.ts
@@ -1,13 +1,10 @@
 import chalk from "chalk";
 import express from "express";
-import { worldMonuments } from "../monuments/data.js";
+
+const port = 3000;
 
 const app = express();
 
-app.listen(3000, () => {
+app.listen(port, () => {
   console.log(chalk.green("Listening on port 3000"));
-});
-
-app.get("/monuments", (_req, res) => {
-  res.json({ worldMonuments });
 });
