+++ src/monument/controller/types.ts
@@ -1,0 +1,5 @@
+import type { Request, Response } from "express";
+
+export interface MonumentsControllerStructure {
+  getMonuments: (req: Request, res: Response) => void;
+}
